Add queue system for rank fetching
Requesting rank for a map the server has no stored leaderboard for
causes the map to be added to a queue for leaderboard fetching.
This allows multiple requests to different missing maps simultaneously.
Maps are handled on a first-come-first-serve basis.
